+++ packages/spot-consumer/src/types/spot-market.ts
@@ -110,7 +110,6 @@ export interface SpotLimitOrderParams {
 
 export interface SpotOrderCancelParams {
   orderHash: string
-  isBuy: boolean
 }
 
 export {